requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
dj-static==0.0.6
Django==3.2.14
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-dotenv==1.4.2
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-webpack-loader==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.4
install==1.3.5
Pillow==9.2.0
PyJWT==2.5.0
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
static3==0.7.0
urllib3==1.26.12
